ml/requirements.txt
numpy==1.26.4
opencv-python==4.9.0.80
mediapipe==0.10.9
psycopg[binary]
pandas
scipy
sqlalchemy
pydantic
fastapi
uvicorn
openai
python-dotenv
tqdm
